// src/api/openapi/model/client.ts
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Address } from './address';
import { Language } from './language';
import { Gender } from './gender';
import { Lock } from './lock';


export interface Client { 
    lock: Lock;
    name: any | null;
    lastname: any | null;
    isCompany: any | null;
    mail1: any | null;
    mail2: any | null;
    tel1: any | null;
    tel2: any | null;
    contact_person: any | null;
    vat_number: any | null;
    fiscal_code: any | null;
    codice_destinatario: any | null;
    pec: any | null;
    esigibilita_iva: any | null;
    publica_amministrazione: any | null;
    cup: any | null;
    cig: any | null;
    id: any | null;
    gender?: Gender;
    address: Address;
    language: Language;
    fullname?: any | null;
    contacts: any | null;
}